pyproject.toml
[tool.poetry]
name = "nonebot_plugin_mc_watcher"
version = "1.0.5"
description = "一个适用于监控多服务器的 Motd 工具。"
license = "GPL3"
authors = ["Lonely-Sails"]
readme = "README.md"
homepage = "https://github.com/Lonely-Sails/nonebot-plugin-mc-watcher"
repository = "https://github.com/Lonely-Sails/nonebot-plugin-mc-watcher"
documentation = "https://github.com/Lonely-Sails/nonebot-plugin-mc-watcher#readme"
keywords = ["nonebot", "nonebot2", "mc", "minecraft", "onebot"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
mcproto = "^0.5.0"
nonebot-plugin-localstore = "^0.7.0"
nonebot-adapter-onebot = "^2.4.3"
pydantic = "^2.7.4"
